pvmfw: smccc: Hack MMIO_GUARD_MAP to expect i32
Until the fix lands in upstream kernels, expect the HVC64 to return a
signed 32-bit integer instead of the SMCCC-compliant i64.

Bug: 253586500
Test: call mmio_guard_map() from non-protected VMs

+++ pvmfw/src/smccc.rs
@@ -113,5 +113,5 @@ pub fn mmio_guard_map(ipa: u64) -> Result<()> {
 
     let res = hvc64(VENDOR_HYP_KVM_MMIO_GUARD_MAP_FUNC_ID, args);
 
-    check_smccc_err(res[0] as i64)
+    check_smccc_err(res[0] as u32 as i32 as i64) // TODO(b/253586500)
 }
